moved styling to new folder
