+++ src/components/plots/styling/AxesDesign.jsx
