+++ src/components/plots/styling/ColorBrowser.jsx
