+++ src/components/plots/styling/ColourInversion.jsx
