+++ src/components/plots/styling/DimensionsRangeEditor.jsx
