+++ src/components/plots/styling/FontDesign.jsx
