+++ src/components/plots/styling/LabelsDesign.jsx
