+++ src/components/plots/styling/LegendEditor.jsx
