+++ src/components/plots/styling/PointDesign.jsx
