+++ src/pages/experiments/[experimentId]/data-processing/filter-cells/components/GenesVsUMIs/GenesVsUMIs.jsx
@@ -14,7 +14,7 @@ import { Vega } from 'react-vega';
 import plot1Pic from '../../../../../../../../static/media/plot7.png';
 import plot2Pic from '../../../../../../../../static/media/plot8.png';
 import plotData from './new_data.json';
-import PlotStyling from '../../../../../../../components/plotStyling';
+import PlotStyling from '../../../../../../../components/plots/styling/PlotStyling';
 import BandwidthOrBinstep from '../ReadAlignment/PlotStyleMisc';
 
 const { Panel } = Collapse;
